# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Express Server API 文档",
    "version": "0.0.1",
    "description": "床前明月光，疑是地上霜。举头望明月，低头思故乡。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1/api"
    }
  ],
  "paths": {
...
    "/blog/tags": {
      "get": {
        "summary": "获取标签列表",
        "description": "分页获取所有标签",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取标签列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tags": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "获取标签列表失败"
          }
        }
      },
      "post": {
        "summary": "创建新标签",
        "description": "创建一个新的文章标签",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "标签名称"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "标签创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "tag": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "创建标签失败"
          }
        }
      }
    },
    "/blog/tags/{id}": {
      "delete": {
        "summary": "删除标签",
        "description": "根据标签ID删除标签",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "标签唯一标识符"
          }
        ],
        "responses": {
          "200": {
            "description": "标签删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "标签未找到"
          },
          "500": {
            "description": "删除标签失败"
          }
        }
      }
    },
...
    "/blog/comments/{articleId}": {
      "post": {
        "summary": "创建新评论",
        "description": "创建一条新的文章评论",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "文章唯一标识符"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "评论内容"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "评论创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "comment": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        }
                      }
                    } 
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "创建评论失败"
          }
        }
      }
    },
    "/blog/comments/{articleId}/{id}": {
      "delete": {
        "summary": "删除评论",
        "description": "根据评论ID删除评论",
        "tags": ["Blog"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "文章唯一标识符"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "评论唯一标识符"
          }
        ],
        "responses": {
          "200": {
            "description": "评论删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "评论未找到"
          },
          "500": {
            "description": "删除评论失败"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}